arch: arm: ipqsoc: align buffer address before flush
This patch aligns the cache boundard address with
respect to the SYS_CACHELINE_SIZE before flushing
them.

SPI driver before initiating the bam transaction
both data and bam desc buffer needs to be cache
flushed. So, we have allocated the desc buffers in
such a manner which is aligned with the cacheline
size.

But few buffers, which are coming from the uboot
framework itself is not aligned. Example during
environment variable save condition. In such cases,
we may need to manually align the buffers address
with respect to the cacheline size.

Change-Id: I40cd3ce21f89047c446d7872189f91a96555440e
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>