board: qti: Fixing AQ firmware loading malloc size
Previously, partition size of the ETHPHYFW is used to
alloc the memory for the loading the AQ firmware. Since
in NAND flash partition, we have 1024K partition size,
which in-turn causes the memory allocation to fail due
to lowmem.

Though it was allocated with 1024K mem, it is not really
needed for AQ firmware loading since the firmware itself
is only 512K. So, fixing the partition size to 512K for
the non-mmc flash types.

Change-Id: Iec78185d07cb91b24f03cf4b16114027c93108d2
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>