drivers: net: phy: update link status ethernet switch
In qca8x8x switch mode, we returned the link status as
always up to MAC layer. But when no port is connected in the
switch, network stack tried to ping the link-parter which is
not at all present in the first place and gets failed eventually.

So, updating the link status as up only when any one port of
switch is up.

Change-Id: I065c9006730b7698e10c083cf5013adfb21ba75c
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>