arch: arm: dts: add support reset SFP TX Pin state
We used to configure SFP TX pin with 8MA drive strenght and pull
up enabled, but later identified that, it is not really required.
Then, removed it in miami.

But, looks likes when device like SFP-80M module is connected to
SFP cage, these configurations are needed. So, enabling it on
all miami sfp rdps.

This change wont affect our SFP-30M module behaviour.

Change-Id: Ieeb7802eb9bf65da1dccdd832f844df648b989bb
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>