board: qti: crashdump: copy then compress the uboot region
Previously, we used to skip the uboot region in the crashdumps.
But, now we dump that one as well. Contents of uboot region may
vary, because uboot is currently running. During compressed
crashdump, when gzip tries to compress the uboot region content
used to calculate crc and content which are actually getting
compressed as data will vary.

So, copy the lower region of RAM (which has the uboot region)
to upper region, then it will become constant. Now, compress
and dump the compressed data.

Change-Id: I2ce826bd7a27f1bec1901ba8cc106763b35f1587
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>