drivers: net: qti: configure mac type in qca8X8X PHY mode
Currently for QCA8X8X PHY mode, we used to update the mac
type as XGMAC only for the 1st port. It causes network
down issue on other ports.

Now, update the driver to configure macs for all the ports.
But, still uniphy needs to be configured only once since it
shared across ports.

Change-Id: Iaa7a9d7e79cb3931e1df9da46f969428a3472b98
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>