driver: mtd: nand: enable runtime layout switch.
This change dynamically de-initializes and initializes
the NAND device and fixes the page size as given below:
1. SBL region --->   2K page size
2. Non-SBL region --->   4K page size

Note: The layout switch is applicable only for
	NAND devices with a 4K page size.

In the IPQ9574 SoC, PBL doesn't support reading on a
4K page layout. Therefore, the SBL firmware should be
written in a 2K page format.

During the writing of a single image on a NAND device
with a 4K page size, the SBL will be written in a 2K page format,
while the rest of the images will be written in a 4K page format.

Change-Id: I209716d2e0f8f2d2ac8ef5458d2e563d2f368fed
Signed-off-by: Gurumoorthy Santhakumar <quic_gsanthak@quicinc.com>