drivers: net: add support to dynamic identify SFP module
This patch helps uboot to identify the presence of SFP
or QCA81XX using I2C bus.

By default, the dts will be configured for QCA81XX module.
U-boot will check for presence of PHY mentioned in the DTS
by probing on the i2c bus. If present, u-boot will proceed,
else it will check for SFP.

If SFP also not found, it ignores that port.

Change-Id: I0c96e4f13ee78d55a3a1015fc50bd9fdb8f4b3df
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>