board: Fix minidump magic read via aligned alloc
The memory allocated using malloc() was not aligned
to the CPU's cache line size, which caused functional
issues such as failure to correctly read the minidump
magic value.

This patch replaces malloc() with malloc_cache_aligned()
to ensure proper alignment, resolving the issue and enabling
reliable access to the minidump data.

Change-Id: I2a9b4b142cecc22fbc7313a69bd463a6c611ee99
Signed-off-by: Poovendhan Selvaraj <poovendh@qti.qualcomm.com>